Clean up README
The README and the docs have `Scenario`s being created manually. I would like the `Scenario` struct to be implemented in such a way that the user never has to really look at it. That is, provide interfaces to create and modify scenarios such that the actual implementation details don't matter.