New scenario builder interface
Made `Scenario`s more modular, they are now a combination of a `Release` and an `Atmosphere` and the `scenario_builder` generates them using a source model and an atmosphere. The only source model is the liquid jet, unchanged from before.

Also changed the simplejet plume model to pre-calculate the density correction.

Adjusted all plume and puff model solutions to no longer depend on the internals of the `Scenario`, the solutions reference themselves and the solution builder unpacks the `Scenario` and is thus the only thing that really needs to know how `Scenario`s work.

TODO: fix the code around `plumerise` to not depend on `Scenrio` internals, have all parameters passed by the solution builder.