Added methods to retrieve properties
Added methods to retrieve properties from the core types (Scenario, etc.) so that I can change the structure of release scenarios, add new things, whatever, and not have to re-write everything. Now when a model needs some property of a scenario it just calls the method, and is agnostic to the internal workings.