Updated correlations and corrections for near and far field
Updated correlations to use the versions from TSCREEN, updated the near-field to handle large \alpha, where the interpolation and near-field overlap, also updated the far-field region to decay like 1/x^2 starting from the last concentration in the interpolation region.

Updated the back end of the plume, instead of guessing at a shape it is just a box since nothing is given in the Britter-McQuaid model.