Testing correlations
Moved Britter-McQuaid correlations to utils, added tests for correlations.

Changed the definition for the vertical extent, I'm fairly sure there is a mistake in the Britter-McQuaid workbook as the vertical extent formula makes no sense -- the plume shrinks rapidly until it is millimeters tall causing mass to disappear out of the system.