Initial SLAB model
An inital version of the SLAB model. Only implements the horizontal jet release parts by implementing each piece entirely as the FORTRAN code runs things.

Uses some structs to clean up the input and output, but the design philosophy is to replicate what SLAB does, but with a nicer interface and in julia.