Re-organizing SLAB initialization and main calculation
The first version was basically the fortran code, lightly edited into julia. This is re-arranged to break out the initializations of different release types into different functions. Also broke out the integration loops into separate functions. This simplifies the main code and makes it easier to un-tangle the goto statements.