Updating unit tests
Tests now compare the output from a SLAB run of the Burro LNG trials with the results from julia.

SLAB returns two sets of output: a table of parameters that can be interpolated to calculate the (time averaged) concentration contours, and a table of pre-calculated results. Annoyingly the pre-calculated results are calculated in a *slightly* different way, though they are presented in the user manual as equivalent. The test data here was calculated based on interpolating the parameters and using equation 56, which was compared to the results calculated by SLAB using equation 55 and the differences validated to be due to SLAB.