More specific structs
Many structs use abstract types inside the struct, e.g.

```julia
struct Example
   a::Number
   b::Number
end
```

A more performant way of doing this (according to the julia manual) is

```julia
struct Example{F<:Number}
    a::F
    b::F
end
```

So I went around and made that change, and added a helper to promote different number types to be consistent with one another, e.g. promot Ints to Floats when a mix of Ints and Floats is being used. I don't know if this will materially change things, but it looks nice.

I also changed the equation sets to dispatch on the same lazy keyword approach as everything else, e.g. using `CCPSRural` instead of `CCPSRural()`. I had initially thought that an equation set could have parameters passed to it, but now I think that not something I want to support. This simplifies the code dramatically since I don't have to handle default behaviour for each equation set, if it doesn't support a correlation type then nothing needs to be written and the defaults handle the `<:Any` case.