Store initial variables at nxi
Initial state of the program is stored at nxi, the start of the integration loop. It will immediately get over-written, allows the initialization routines to pass the initial state to the integration routines without using a lot of extra variables.

The original fortran code keeps the entire state as a bunch of global variables, which I would rather not.