DataInterpolations v8.0.1 compat
A change in DataInterpolations v8.0.1 enforces the ordering of the independent variable, has to be monotonically increasing. Apparently, in one test, SLAB returned a backwards timestep. I'm not sure how this is possible, but for safety ensure that the x and t correlations are sorted to be in ascending order.